requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
fastapi==0.104.1
Flask==3.0.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
pandas==2.1.3
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
Werkzeug==3.0.1
